refactor(qwk): split Phase 3 dedup/validation tests into own file
